Robustify `StreamExt::ToMemoryStream`
Previously, if an error occured while the stream was being copied
into memory, the open Stream would leak, even if the method was
employed in a way that would transfer ownership of the stream into it.

The `closeStream` parameter has additionally been renamed to
`takeOwnership` to reflect the semantics more clearly.